Code running enabled
This is the final version implementing a virtual machine to simulate a code runner for Little Duck. The main includes the parser and intermediate code generator. To run main.py you need a source code in Little Duck. You can edit the name of the file to be read in the bottom part of the main section. This source code in Little Duck tests several capabilities of the compiler. When running main.py it will generate a .obj file with the intermediate code that the virtual machine simulator reads and runs.